+++ lib/forest_liana/bootstrapper.rb
@@ -76,7 +76,11 @@ module ForestLiana
     end
 
     def fetch_models
-      ActiveRecord::Base.subclasses.each { |model| fetch_model(model) }
+      if Rails::VERSION::MAJOR > 5
+        ApplicationRecord.subclasses.each { |model| fetch_model(model) }
+      else
+        ActiveRecord::Base.subclasses.each { |model| fetch_model(model) }
+      end
     end
 
     def fetch_model(model)
@@ -88,6 +92,7 @@ module ForestLiana
             model.descendants.each { |submodel_sti| fetch_model(submodel_sti) }
           end
 
+          model.connection
           if analyze_model?(model)
             ForestLiana.models << model
           end